+++ spec/flow/operation/accessors_spec.rb
@@ -4,17 +4,22 @@ RSpec.describe Flow::Operation::Accessors, type: :module do
   include_context "with an example operation"
 
   let(:operation_class) { example_operation_class }
-  let(:state_attribute) { Faker::Lorem.word.to_sym }
+  let(:state_attribute) { Faker::Lorem.unique.word.to_sym }
+  let(:state_attributes) { [ state_attribute ] }
   let(:state_attribute_writer) { "#{state_attribute}=".to_sym }
   let(:state_attribute_value) { Faker::Hipster.word }
   let(:definition_options) { {} }
+  let(:prefix) { false }
 
   shared_context "with prefix option set" do
     let(:definition_options) { { prefix: prefix } }
     let(:prefix_name) { (prefix == true ? :state : prefix).to_sym }
   end
 
-  before { example_state_class.attr_accessor(state_attribute) }
+  shared_context "when multiple method names are given" do
+    let(:other_state_attribute) { Faker::Lorem.unique.word.to_sym }
+    let(:state_attributes) { [ state_attribute, other_state_attribute ] }
+  end
 
   shared_examples_for "it has exactly ? of type" do |count, tracker_type|
     subject { operation.public_send(tracker_name).count(state_attribute) }
@@ -67,9 +72,33 @@ RSpec.describe Flow::Operation::Accessors, type: :module do
     end
   end
 
+  shared_examples_for "it defines delegated readers for all methods" do
+    it "defines readers for each method" do
+      state_attributes.each do |attr|
+        if prefix
+          expect(operation).to delegate_method(attr).to(:state).with_prefix(prefix_name)
+        else
+          expect(operation).to delegate_method(attr).to(:state)
+        end
+      end
+    end
+  end
+
+  shared_examples_for "it defines delegated writers for all methods" do
+    it "defines writers for each method" do
+      state_attributes.each do |attr|
+        if prefix
+          expect(operation).to delegate_method("#{attr}=".to_sym).to(:state).with_prefix(prefix_name).with_arguments(state_attribute_value)
+        else
+          expect(operation).to delegate_method("#{attr}=".to_sym).to(:state).with_arguments(state_attribute_value)
+        end
+      end
+    end
+  end
+
   describe ".state_reader" do
     subject(:operation) do
-      operation_class.__send__(:state_reader, state_attribute, **definition_options)
+      operation_class.__send__(:state_reader, *state_attributes, **definition_options)
       operation_class.new(example_state)
     end
 
@@ -80,6 +109,12 @@ RSpec.describe Flow::Operation::Accessors, type: :module do
     it_behaves_like "it has exactly one tracker variable of type", :reader
     it_behaves_like "it has no tracker variables of type", :accessor
 
+    context "when multiple method names are given" do
+      include_context "when multiple method names are given"
+
+      it_behaves_like "it defines delegated readers for all methods"
+    end
+
     context "when prefix is given" do
       include_context "with prefix option set"
 
@@ -89,6 +124,12 @@ RSpec.describe Flow::Operation::Accessors, type: :module do
         it_behaves_like "it has exactly one tracker variable of type", :reader
 
         it { is_expected.to delegate_method(state_attribute).to(:state).with_prefix(prefix_name) }
+
+        context "when multiple method names are given" do
+          include_context "when multiple method names are given"
+
+          it_behaves_like "it defines delegated readers for all methods"
+        end
       end
 
       context "when prefix is a string" do
@@ -97,6 +138,12 @@ RSpec.describe Flow::Operation::Accessors, type: :module do
         it_behaves_like "it has exactly one tracker variable of type", :reader
 
         it { is_expected.to delegate_method(state_attribute).to(:state).with_prefix(prefix_name) }
+
+        context "when multiple method names are given" do
+          include_context "when multiple method names are given"
+
+          it_behaves_like "it defines delegated readers for all methods"
+        end
       end
     end
 
@@ -117,7 +164,7 @@ RSpec.describe Flow::Operation::Accessors, type: :module do
 
   describe ".state_writer" do
     subject(:operation) do
-      operation_class.__send__(:state_writer, state_attribute, **definition_options)
+      operation_class.__send__(:state_writer, *state_attributes, **definition_options)
       operation_class.new(example_state)
     end
 
@@ -128,6 +175,12 @@ RSpec.describe Flow::Operation::Accessors, type: :module do
     it_behaves_like "it has exactly one tracker variable of type", :writer
     it_behaves_like "it has no tracker variables of type", :accessor
 
+    context "when multiple method names are given" do
+      include_context "when multiple method names are given"
+
+      it_behaves_like "it defines delegated writers for all methods"
+    end
+
     context "when prefix is given" do
       include_context "with prefix option set"
 
@@ -137,6 +190,12 @@ RSpec.describe Flow::Operation::Accessors, type: :module do
         it_behaves_like "it has exactly one tracker variable of type", :writer
 
         it { is_expected.to delegate_method(state_attribute_writer).to(:state).with_prefix(prefix_name).with_arguments(state_attribute_value) }
+
+        context "when multiple method names are given" do
+          include_context "when multiple method names are given"
+
+          it_behaves_like "it defines delegated writers for all methods"
+        end
       end
 
       context "when prefix is a string" do
@@ -145,6 +204,12 @@ RSpec.describe Flow::Operation::Accessors, type: :module do
         it_behaves_like "it has exactly one tracker variable of type", :writer
 
         it { is_expected.to delegate_method(state_attribute_writer).to(:state).with_prefix(prefix_name).with_arguments(state_attribute_value) }
+
+        context "when multiple method names are given" do
+          include_context "when multiple method names are given"
+
+          it_behaves_like "it defines delegated writers for all methods"
+        end
       end
     end
 
@@ -165,7 +230,7 @@ RSpec.describe Flow::Operation::Accessors, type: :module do
 
   describe ".state_accessor" do
     subject(:operation) do
-      operation_class.__send__(:state_accessor, state_attribute, **definition_options)
+      operation_class.__send__(:state_accessor, *state_attributes, **definition_options)
       operation_class.new(example_state)
     end
 
@@ -177,6 +242,13 @@ RSpec.describe Flow::Operation::Accessors, type: :module do
     it_behaves_like "it has exactly one tracker variable of type", :writer
     it_behaves_like "it has exactly one tracker variable of type", :reader
     it_behaves_like "it has exactly one tracker variable of type", :accessor
+
+    context "when multiple method names are given" do
+      include_context "when multiple method names are given"
+
+      it_behaves_like "it defines delegated readers for all methods"
+      it_behaves_like "it defines delegated writers for all methods"
+    end
 
     context "when prefix is given" do
       include_context "with prefix option set"
@@ -190,6 +262,13 @@ RSpec.describe Flow::Operation::Accessors, type: :module do
 
         it { is_expected.to delegate_method(state_attribute).to(:state).with_prefix(prefix_name) }
         it { is_expected.to delegate_method(state_attribute_writer).to(:state).with_prefix(prefix_name).with_arguments(state_attribute_value) }
+
+        context "when multiple method names are given" do
+          include_context "when multiple method names are given"
+
+          it_behaves_like "it defines delegated readers for all methods"
+          it_behaves_like "it defines delegated writers for all methods"
+        end
       end
 
       context "when prefix is a string" do
@@ -201,6 +280,13 @@ RSpec.describe Flow::Operation::Accessors, type: :module do
 
         it { is_expected.to delegate_method(state_attribute).to(:state).with_prefix(prefix_name) }
         it { is_expected.to delegate_method(state_attribute_writer).to(:state).with_prefix(prefix_name).with_arguments(state_attribute_value) }
+
+        context "when multiple method names are given" do
+          include_context "when multiple method names are given"
+
+          it_behaves_like "it defines delegated readers for all methods"
+          it_behaves_like "it defines delegated writers for all methods"
+        end
       end
     end